Add header function and prompt for the fish shell
The `soji header` function isn't compatible with the fish shell. It's
pretty typical in fish to encapsulate that kind of thing as a standalone
function which can be called from a prompt function, so that's what I've
done here. It should look very similar to `soji header`.

I wasn't sure where to put it so I created a directory in the root.

https://fishshell.com/