Remove all crontab entries (#20)
* Remove all crontab entries

Before starting a new log entry, remove current and past crontab
entries. This will allow for finishing tasks early as well as keep the
crontab file cleaner.

* Add header function and prompt for the fish shell (#19)

The `soji header` function isn't compatible with the fish shell. It's
pretty typical in fish to encapsulate that kind of thing as a standalone
function which can be called from a prompt function, so that's what I've
done here. It should look very similar to `soji header`.

I wasn't sure where to put it so I created a directory in the root.

https://fishshell.com/